Fix develop mode without TS and refacto strapi-admin file
